+++ src/Battery/Utils/StringUtils.cpp
@@ -172,7 +172,7 @@ namespace Battery {
 			char_array_4[i++] = data[in_]; in_++;
 			if (i == 4) {
 				for (i = 0; i < 4; i++)
-					char_array_4[i] = base64_chars.find(char_array_4[i]);
+					char_array_4[i] = (uint8_t)base64_chars.find(char_array_4[i]);
 
 				char_array_3[0] = (char_array_4[0] << 2) + ((char_array_4[1] & 0x30) >> 4);
 				char_array_3[1] = ((char_array_4[1] & 0xf) << 4) + ((char_array_4[2] & 0x3c) >> 2);
@@ -189,7 +189,7 @@ namespace Battery {
 				char_array_4[j] = 0;
 
 			for (j = 0; j < 4; j++)
-				char_array_4[j] = base64_chars.find(char_array_4[j]);
+				char_array_4[j] = (uint8_t)base64_chars.find(char_array_4[j]);
 
 			char_array_3[0] = (char_array_4[0] << 2) + ((char_array_4[1] & 0x30) >> 4);
 			char_array_3[1] = ((char_array_4[1] & 0xf) << 4) + ((char_array_4[2] & 0x3c) >> 2);
